% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT2.R
\name{CIBERSORT2}
\alias{CIBERSORT2}
\title{Main functions}
\usage{
CIBERSORT2(LM22_matrix, mixture, opt_path, perm = 0, QN = TRUE)
}
\arguments{
\item{LM22_matrix}{Gene expression matrix from isolated cells (LM22)}

\item{mixture}{A data.frame or matrix.heterogenous mixed expression}

\item{opt_path}{The output path of the calculation result.
If the path does not contain the ".txt" file name, the output file name is "CIBERSORT-Results.txt".}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
Main functions
}
